pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comment-llm"
version = "0.1.0"
description = "A RAG-based system for analyzing Google Maps reviews"
authors = [{name = "Comment-LLM", email = ""}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "selenium>=4.15.2",
    "beautifulsoup4>=4.12.2",
    "requests>=2.31.0",
    "pandas>=2.1.3",
    "numpy>=1.25.2",
    "chromadb>=0.4.18",
    "openai>=1.3.7",
    "python-dotenv>=1.0.0",
    "langchain>=0.0.340",
    "langchain-openai>=0.0.2",
    "sentence-transformers>=2.2.2",
    "click>=8.1.7",
    "tqdm>=4.66.1",
]

[project.scripts]
comment-llm = "comment_llm.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["comment_llm*"]